requirements.txt
attrs==20.3.0
Brotli==1.0.9
click==7.1.2
coverage==5.4
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
packaging==20.9
pandas==1.2.1
patsy==0.5.1
plotly==4.14.3
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scipy==1.6.0
six==1.15.0
statsmodels==0.12.2
toml==0.10.2
waitress==1.4.4
Werkzeug==1.0.1
